refactor map client to use dedicated hook and services
